requirements.txt
Flask
gunicorn
bcrypt
flask_mail
psycopg2-binary
python-dotenv
